+++ R/rowFiltering.R
@@ -1,12 +1,16 @@
 #' Standard deviation outlier filtering
 #'
 #' Remove outliers based on standard deviation thresholds. \cr
-#' Only values within \code{mean - sd * n_sigmas} and \code{mean + sd * n_sigmas} are kept
+#' Only values within \code{mean - sd * n_sigmas} and \code{mean + sd * n_sigmas} are kept.
 #' @param dataSet Matrix, data.frame or data.table
 #' @param cols List of numeric column(s) name(s) of dataSet to transform. To transform all 
 #' numeric columns, set it to "auto".  (character, default to "auto")
 #' @param n_sigmas number of times standard deviation is accepted (interger, default to 3)
 #' @param verbose Should the algorithm talk? (logical, default to TRUE)
+#' @details Filtering is made column by column, meaning that extrem values from first element
+#' of \code{cols} are removed, then extrem values from second element of \code{cols} are removed, 
+#' ... \cr
+#' So if filtering is perfomed on too many column, there ia high risk that a lot of rows will be dropped.
 #' @return Same dataset with less rows, edited by \strong{reference}. \cr
 #' If you don't want to edit by reference please provide set \code{dataSet = copy(dataSet)}.
 #' @examples 
@@ -71,6 +75,10 @@ remove_sd_outlier <- function(dataSet, cols = "auto", n_sigmas = 3, verbose = TR
 #' columns, set it to "auto".  (character, default to "auto")
 #' @param threshold share of occurencies under which row should be removed (numeric, default to 0.01)
 #' @param verbose Should the algorithm talk? (logical, default to TRUE)
+#' @details Filtering is made column by column, meaning that extrem values from first element
+#' of \code{cols} are removed, then extrem values from second element of \code{cols} are removed, 
+#' ... \cr
+#' So if filtering is perfomed on too many column, there ia high risk that a lot of rows will be dropped.
 #' @return Same dataset with less rows, edited by \strong{reference}. \cr
 #' If you don't want to edit by reference please provide set \code{dataSet = copy(dataSet)}.
 #' @examples 
@@ -122,5 +130,71 @@ remove_rare_categorical <- function(dataSet, cols ="auto", threshold = 0.01, ver
   }
   
   ## Wrap-up
+  return(dataSet)
+}
+
+#' Percentile outlier filtering
+#'
+#' Remove outliers based on percentiles. \cr
+#' Only values within \code{n}th and \code{100 - n}th percentiles are kept.
+#' @param dataSet Matrix, data.frame or data.table
+#' @param cols List of numeric column(s) name(s) of dataSet to transform. To transform all 
+#' numeric columns, set it to "auto".  (character, default to "auto")
+#' @param percentile percentiles to filter (numeric, default to 1)
+#' @param verbose Should the algorithm talk? (logical, default to TRUE)
+#' @details Filtering is made column by column, meaning that extrem values from first element
+#' of \code{cols} are removed, then extrem values from second element of \code{cols} are removed, 
+#' ... \cr
+#' So if filtering is perfomed on too many column, there ia high risk that a lot of rows will be dropped.
+#' @return Same dataset with less rows, edited by \strong{reference}. \cr
+#' If you don't want to edit by reference please provide set \code{dataSet = copy(dataSet)}.
+#' @examples 
+#' # Given
+#' library(data.table)
+#' dataSet <- data.table(num_col = 1:100)
+#' 
+#' # When
+#' dataSet <- remove_percentile_outlier(dataSet, cols = "auto", percentile = 1, verbose = TRUE)
+#' 
+#' # Then extrem value is no longer in set
+#' 1 %in% dataSet[["num_col"]] # Is false
+#' 2 %in% dataSet[["num_col"]] # Is true
+#' @importFrom stats quantile
+#' @export
+remove_percentile_outlier <- function(dataSet, cols = "auto", percentile = 1, verbose = TRUE){
+  ## Environement
+  function_name <- "remove_percentile_outlier"  
+  
+  ## Sanity check
+  dataSet <- checkAndReturnDataTable(dataSet = dataSet)
+  cols <- real_cols(dataSet = dataSet, cols = cols, function_name = function_name, types = "numeric")
+  
+  ## Initialization
+  if (verbose){
+    printl(function_name, ": I start to filter categorical rare events")
+    pb <- initPB(function_name, names(dataSet))
+    start_time <- proc.time()
+  }
+  initial_nrow <- nrow(dataSet)
+  
+  ## Computation
+  for (col in cols){
+    tmp_nrow <- nrow(dataSet)
+    percentiles <- quantile(dataSet[[col]], 
+                            c(percentile / 100, (100 - percentile) / 100), na.rm = TRUE)
+    dataSet <- dataSet[(get(col) >=  percentiles[1]) & (get(col) <= percentiles[2]), ]
+    if (verbose){
+      printl(function_name, ": dropped ", tmp_nrow - nrow(dataSet), " row(s) that are rare event on ", col, ".")
+      setPB(pb, col)
+    }
+  }
+  
+  if (verbose){
+    printl(function_name, ": ", initial_nrow - nrow(dataSet), " have been dropped. It took ", 
+           round( (proc.time() - start_time)[[3]], 2), " seconds. ")
+  }
+  
+  
+  ## Wrapp-up
   return(dataSet)
 }